# public/swagger.json
{
	"components": {
...
		"schemas": {
			"ResponseData": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [0, 1],
						"description": "1 = succeed | 0 = failed"
					},
					"data": {},
					"messages": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Error/warning messages"
					}
				},
				"required": ["status", "data", "messages"],
				"type": "object",
				"additionalProperties": false
			},
...
			"ObjectID": {
				"description": "Create a new ObjectID instance.",
				"properties": {
					"generationTime": {
						"type": "number",
						"format": "double",
						"description": "The generation time of this ObjectId instance."
					}
				},
				"required": ["generationTime"],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProviderInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"access_token": {
						"type": "string"
					}
				},
				"required": ["name"],
				"type": "object",
				"additionalProperties": false
			},
...
			"User": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string",
						"description": "User name"
					},
					"username": {
						"type": "string",
						"description": "Unique username of a user\nThis equavilent with \"slug\""
					},
					"type": {
						"type": "string",
						"enum": ["user", "service_account"],
						"description": "Service Account is User!"
					},
					"email": {
						"type": "string",
						"description": "User email address"
					},
					"verified": {
						"type": "boolean",
						"description": "Is this user's email or phone verified?"
					},
					"image": {
						"type": "string",
						"description": "User profile picture URL"
					},
					"providers": {
						"items": {
							"$ref": "#/components/schemas/ProviderInfo"
						},
						"type": "array",
						"description": "List of Cloud Providers which this user can access to"
					},
					"password": {
						"type": "string",
						"description": "User password (hashed)"
					},
					"token": {
						"$ref": "#/components/schemas/AccessTokenInfo",
						"description": "User token"
					},
					"roles": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Role"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "User's roles\nTODO: Thêm \"workspace\" cho \"roles\" vì mỗi workspace sẽ có roles khác nhau, trong khi mỗi user có thể nằm trong nhiều workspace"
					},
					"teams": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Team"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "User's team IDs which this user is a member"
					},
					"workspaces": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Workspace"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "List of workspace IDs which this user is a member"
					},
					"activeWorkspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "Active workspace of a user"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Project": {
				"description": "Projects",
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"apiKey": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"lastUpdatedBy": {
						"type": "string"
					},
					"latestBuild": {
						"type": "string"
					},
					"appSlugs": {
						"type": "string"
					},
					"apps": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/ObjectID"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/App"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						],
						"description": "List of App IDs"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"CloudProviderType": {
				"type": "string",
				"enum": ["gcloud", "digitalocean", "custom"]
			},
...
			"CloudProvider": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string",
						"description": "Cloud provider name"
					},
					"shortName": {
						"$ref": "#/components/schemas/CloudProviderType",
						"description": "Cloud provider short name, without spacing & special characters"
					},
					"apiAccessToken": {
						"type": "string",
						"description": "Content of the API access token to use services on this cloud provider\n- Apply for: Digital Ocean"
					},
					"serviceAccount": {
						"type": "string",
						"description": "Content of the Service Account credentials ti access services on this cloud provider\n- Apply for: Google Cloud, AWS,...\n- For example: Kubernetes Clusters, Single Sign-On,..."
					},
					"clusters": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Cluster"
								},
								"type": "array"
							}
						],
						"description": "List of available clusters on this provider"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GitProviderType": {
				"type": "string",
				"enum": ["bitbucket", "github", "gitlab"]
			},
			"GitProvider": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"gitWorkspace": {
						"type": "string"
					},
					"repo": {
						"properties": {
							"sshPrefix": {
								"type": "string"
							},
							"url": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"type": {
						"$ref": "#/components/schemas/GitProviderType"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"project": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Project"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the project"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_GitProvider.keyofHiddenBodyKeys_": {
				"$ref": "#/components/schemas/Pick_GitProvider.Exclude_keyofGitProvider.keyofHiddenBodyKeys__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
			"Omit_CloudProvider.keyofHiddenBodyKeys_": {
				"$ref": "#/components/schemas/Pick_CloudProvider.Exclude_keyofCloudProvider.keyofHiddenBodyKeys__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "API_ACCESS_TOKEN",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "@topgroup/diginext",
		"version": "2.6.5",
		"description": "A BUILD SERVER & CLI to deploy apps to any Kubernetes clusters.",
		"license": {
			"name": "GPL"
		},
		"contact": {
			"name": "TOP GROUP (a.k.a Digitop)",
			"email": "dev@wearetopgroup.com",
			"url": "https://github.com/digitopvn/diginext"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/git": {
			"get": {
				"operationId": "Read",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/GitProvider"
														},
														{
															"items": {
																"$ref": "#/components/schemas/GitProvider"
															},
															"type": "array"
														}
													]
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Find one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Mark this request as search (return the similar results based on the filter query params)",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"default": "true",
							"type": "boolean"
						}
					},
					{
						"description": "If `true`, return the excel binary file to download.",
						"in": "query",
						"name": "download",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/GitProvider"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_GitProvider.keyofHiddenBodyKeys_"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/GitProvider"
													},
													"type": "array"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_GitProvider.keyofHiddenBodyKeys_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"n": {
															"type": "number",
															"format": "double"
														},
														"ok": {
															"type": "number",
															"format": "double"
														}
													},
													"type": "object"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Delete one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Delete one item by `{slug}`",
						"in": "query",
						"name": "slug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/git/ssh/public-key": {
			"get": {
				"operationId": "GetPublicKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Find one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Mark this request as search (return the similar results based on the filter query params)",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"default": "true",
							"type": "boolean"
						}
					},
					{
						"description": "If `true`, return the excel binary file to download.",
						"in": "query",
						"name": "download",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				]
			}
		},
		"/git/ssh/create": {
			"post": {
				"operationId": "CreateKeysSSH",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"publicKey": {
										"type": "string"
									},
									"privateKey": {
										"type": "string"
									}
								},
								"required": ["publicKey", "privateKey"],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/git/ssh/generate": {
			"post": {
				"operationId": "GenerateSSH",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"publicKey": {
													"type": "string"
												}
											},
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/git/ssh/verify": {
			"post": {
				"operationId": "VerifySSH",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": ["Git Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "provider",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/provider": {
			"get": {
				"operationId": "Read",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/CloudProvider"
														},
														{
															"items": {
																"$ref": "#/components/schemas/CloudProvider"
															},
															"type": "array"
														}
													]
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Cloud Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Find one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Mark this request as search (return the similar results based on the filter query params)",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"default": "true",
							"type": "boolean"
						}
					},
					{
						"description": "If `true`, return the excel binary file to download.",
						"in": "query",
						"name": "download",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/CloudProvider"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Cloud Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_CloudProvider.keyofHiddenBodyKeys_"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/CloudProvider"
													},
													"type": "array"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Cloud Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_CloudProvider.keyofHiddenBodyKeys_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"n": {
															"type": "number",
															"format": "double"
														},
														"ok": {
															"type": "number",
															"format": "double"
														}
													},
													"type": "object"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Cloud Provider"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Delete one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Delete one item by `{slug}`",
						"in": "query",
						"name": "slug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}